src/app/components/start/start.component.ts
import { Component, OnInit } from '@angular/core';
import { FormGroup, FormBuilder, FormControl, Validators } from '@angular/forms';

@Component({
  selector: 'app-start',
  templateUrl: './start.component.html',
  styleUrls: ['./start.component.css']
})
export class StartComponent implements OnInit {

  public nomesForm: FormGroup;

  constructor(
    private fb: FormBuilder
  ) { }

  ngOnInit() {
    this.nomesForm = this.fb.group({
      jogador1: new FormControl('', Validators.required),
      jogador2: new FormControl('', Validators.required)
    })
  }

}
